isolated components and made them static
