--- src/components/bs-list.jsx
@@ -1,25 +1,0 @@
-import React, { Component } from 'react';
-import BSListItem from './bs-list-item';
-
-export default class List extends Component{
-    render(){
-        let config = this.props.config;
-        let listItems = this.props.links.map(function(link, n){
-            return <BSListItem key={n} link={link} config={config}/>
-        });
-        if(this.props.rootNav){
-            return(
-                <ul className='nav navbar-nav'>
-                    {listItems}
-                </ul>
-            );
-        }
-        return(
-            <ul className='dropdown-menu'>
-                {listItems}
-            </ul>
-        );
-        
-    }
-    
-}
